plat/arm: sds: Move to drivers/ folder
