+++ plat/arm/board/juno/juno_bl1_setup.c
@@ -9,12 +9,11 @@
 #include <common/bl_common.h>
 #include <common/debug.h>
 #include <common/tbbr/tbbr_img_def.h>
+#include <drivers/arm/css/sds.h>
 #include <drivers/arm/sp805.h>
 #include <plat/arm/common/plat_arm.h>
 #include <plat/common/platform.h>
 #include <platform_def.h>
-
-#include <sds.h>
 
 void juno_reset_to_aarch32_state(void);
